refactor: ReservationManager 클래스 패키지 변경 및 import 정리
